+++ src/lib/index.ts
@@ -8,29 +8,42 @@ export interface Validator<T> {
   validate(state: T): boolean;
 }
 
-interface Config<T extends string> {
+type ReturnType = "object" | "array";
+
+type ResultValue = Error[] | { [id: string]: Error };
+
+interface Config<T extends string, R extends ReturnType> {
   errorStateId?: T;
+  returnType?: R;
 }
-interface IntenalParams<T extends string> extends Config<T> {
-  results: Error[];
+
+interface IntenalParams<T extends ResultValue> {
+  results: T;
 }
 
 interface Reducer<State extends {}, Action> {
   (state: State, action: Action): State;
 }
 
+const defaultConfig = {
+  errorStateId: "errors",
+  returnType: "object"
+};
+
+type WithValidatorState<State, Key extends string, ReturnType> = State &
+  Record<Key, ReturnType extends "object" ? { [id: string]: Error } : Error[]>;
+
 class ValidationWatcher {
   public get watchRootReducer() {
     return this._watchRootReducer;
   }
 
-  private internal: IntenalParams<string>;
+  private internal: IntenalParams<ResultValue>;
 
-  constructor(state = {}) {
+  constructor(internal = {}) {
     this.internal = {
-      errorStateId: "errors",
       results: [],
-      ...state
+      ...internal
     };
   }
 
@@ -63,32 +76,62 @@ class ValidationWatcher {
   private _watchRootReducer = <
     State extends {},
     Action,
-    Key extends string = "errors"
+    Key extends string = "errors",
+    TReturnType extends ReturnType = "object"
   >(
     reducer: Reducer<State, Action>,
-    config: Config<Key> = {}
-  ): Reducer<State & Record<Key, Error[]>, Action> => {
-    const { errorStateId } = config;
-    if (errorStateId) {
-      this.internal.errorStateId = errorStateId;
+    config: Config<Key, TReturnType> = {}
+  ): Reducer<WithValidatorState<State, Key, TReturnType>, Action> => {
+    const { errorStateId, returnType }: { [id: string]: string } = {
+      ...defaultConfig,
+      ...config
+    };
+    if (returnType === "object") {
+      this.internal.results = {};
+    } else {
+      this.internal.results = [];
     }
     return (state, action) => {
       if (state && errorStateId in state) {
         delete state[errorStateId];
       }
       const nextState = reducer(state, action);
-      const nextErrors = this.internal.results.slice();
-      this.internal.results = [];
+      const _nextErrors = this.nextErrors();
       return {
         ...(nextState as object),
-        [this.internal.errorStateId]: nextErrors
-      } as State & Record<Key, Error[]>;
+        [errorStateId]: _nextErrors
+      } as State & WithValidatorState<State, Key, TReturnType>;
     };
   };
 
-  private withError(error: Error): void {
-    this.internal.results.push(error);
+  private withError = (error: Error): void => {
+    if (isResultsObject(this.internal.results)) {
+      this.internal.results = {
+        ...this.internal.results,
+        [error.id]: error
+      };
+    } else {
+      this.internal.results = [...this.internal.results, error];
+    }
+  };
+
+  private nextErrors = (): ResultValue => {
+    if (isResultsObject(this.internal.results)) {
+      const _result = Object.assign({}, this.internal.results);
+      this.internal.results = {};
+      return _result;
+    }
+    const _result = [...this.internal.results];
+    this.internal.results = [];
+    return _result;
+  };
+}
+
+function isResultsObject(results): results is { [id: string]: Error } {
+  if (Array.isArray(results)) {
+    return false;
   }
+  return true;
 }
 
 export function getInstance() {